[feat] #10 dto 변경

+++ data/src/main/java/com/dpm/sixpack/data/source/remote/dto/request/FinishRunningRequestDto.kt
@@ -8,9 +8,9 @@ import kotlinx.serialization.Serializable
 @Serializable
 data class FinishRunningRequestDto(
     @SerialName("totalDistance")
-    val totalDistance: Long,
+    val totalDistance: Double,
     @SerialName("totalDuration")
-    val totalDuration: Long,
+    val totalDuration: Int,
     @SerialName("avgPace")
     val avgPace: Int,
     @SerialName("avgCadence")
